requirements.txt
streamlit==1.41.1
openai==1.59.7
crewai==0.100.0
langchain==0.3
langchain-openai==0.2.0
Pillow==11.0.0